fix: correct skiplist macros types and move structs to header
